fix(shared-data): Small cosmetic labware definition fixes (#17868)
This fixes a couple of small bugs in labware definition metadata.

* Fixing the well groups in `nest_96_wellplate_2ml_deep`: Closes
EXEC-1159. This makes the labware display less weirdly in Labware
Library. It was displaying as if the wells had their own brand, like
they were tubes in a tube rack.
* Fixing the well bottom shape in
`usascientific_96_wellplate_2.4ml_deep`: Closes PLAT-144.

Both of the definitions are updated in-place, since we haven't released
them yet.