feat(api): add volume configuration and LPD to new transfers (#17800)
Closes RQA-3941, AUTH-1427, AUTH-1428, AUTH-1296

# Overview

**This PR fixes all the above bugs/ testing feedback requests by adding
the following features:**
1. adds liquid probing before aspirating from source wells when
`liquid_presence_detection` flag is turned on for the pipette
- `transfer_liquid` and `distribute_liquid` can do liquid probing as long as `new_tip` is not
 `never`; (probing cannot succeed with used tips)
- `consolidate_liquid` will NOT do probing for any tip options. (I am
re-thinking this. Original thought was that we can't guarantee that all
src wells would use a new tip but they can if the src volume is so large
that only one src vol can fit into one src-to-dest transfer, or if
volume is larger than tip max volume. In these cases we can guarantee
new tip per source well when `new_tip` is `always` or `per source`, and
can carry out liquid probing)
- a source well is probed only once. So if we revisit the same source
well again, usually because the volume to transfer is larger than tip
volume, we will skip probing after the first visit.
2. calls `configure_for_volume` before a transfer step + makes sure
pipette is 'ready to aspirate' before the aspirate step
- we do this before submerging because a `prepare_to_aspirate` needs to
be called immediately after configuration, and that moves the plunger.
So we need to be above the liquid. This will happen after liquid probing
if LPD is enabled
- the volume we consider for configuration is the last volume we will be
dispensing from that volume in the tip. That is because volume
configuration is only needed when we need to do a push out. So if we are
doing a multi-dispense sequence, only the final dispense will be
performing a push-out. So we use this volume for doing the
configuration.
3. checks and raises errors if submerge start point or retract end point
are inside liquid

**To implement all of this, the order of pre-aspirate submerge steps
needed to be changed as such-**
1. If we are at a source well and need to do volume configuration then-
- move to the probing start point (which is hard coded as 2mm above the
well)
  - remove air gap (probing requires plunger at bottom position)
- if probing should happen, perform probing (which moves the pipette
down until it finds liquid, then it moves back to 2mm above the well).
If probing shouldn't happen, skip to next step
  - configure for volume
  - prepare for aspirate
- check if submerge start point is inside liquid or below aspirate
point, raise errors if it is
  - if not, move to submerge start point
  - start submerging at submerge speed
2. If we are NOT at a source well or we DO NOT need to do volume
configuration (which will be the case for non-first aspiration during
consolidate) then-
- check if submerge start point is inside liquid or below aspirate
point, raise errors if it is
  - if not, move to submerge start point
  - remove any air gap
  - start submerging at submerge speed

## Risk assessment

Medium. Has high impact on pipette prep part of transfer.