feat(api): return FlexStackerShuttleMissing from stacker engine commands (#17881)
# Overview
We want to enable error recovery during a protocol if the Flex Stacker
shuttle is not detected. This PR does the first step, which is to catch
this error. The hardware controller layer now checks for the platform
status before executing dispense and store labware, and raises the
appropriate error.

closes EXEC-1287, EXEC-1200