fix(app): Fix LPC not POSTing labware defs to the maintenance run (#17894)
Closes RQA-4010

Currently, LPC utilizes the existing labware definitions from runs/:runId/loaded_labware_defintions and POSTs those definitions to the LPC maintenance run. However, it turns out that those definitions do not necessarily exist on the run resource at the time of LPC maintenance run creation.

We actually have access to the labware definitions from analysis directly and don't need to query the server for them. A much simpler (and functional) approach is to just POST the definitions we already have.