chore(api,shared-data): update evotip definitions (#17907)
Update the labware definitions for the evotip adapters. These are the
existing geometry for the 4 v2 adapters; they're getting new names (and
thus going to v1), and they're getting their branding details filled in.
We've removed the v1 adapters entirely, since their names were
confusingly close to the new ones.

This also adds the adapters to the labware tab and to the labware
library. Note that it does _not_ add them to PD; PD can't create evotip
protocols, and anyway we only want people to be using these with one
specific protocol.



Closes EXEC-1378
Closes EXEC-1364