fix(app): Fix LPC offset tables not in sync (#17917)
loses RQA-4013

There are a couple pieces of LPC state that are not necessarily controlled by the active app: the deck config and the labware info (which contains offset data). I recently moved the logic for updating store changes whenever the deck config changed to a separate hook, but I didn't do the same for the labware info, woops. This means that if a diferent user was actively performing LPC, offset updates wouldn't reflect in protocol setup offset table.

This commit creates granular state update actions for both the deck config and labware info.