refactor(app): Drive device reset settings fully from client (#17921)
## Overview

This refactors the desktop app and ODD components for doing a device
reset.

Goes towards EXEC-1281.

## Background

Historically, all of the reset options were server-defined. The client
would fetch the available options from the server (via `GET
/settings/reset/options`) and display them as-is, with server-defined
text.

More recently, that model has been stretched beyond its limits:

* We want the UI to be localized with frontend-defined strings.
* We want frontend to control the options' ordering and grouping.
* We want some reset options, like `onDeviceDisplay`, to be special:
they should be hidden from the UI and selected implicitly depending on
the user's selection of other options.
* We want the server to implement some of these reset options through
separate endpoints, instead of all of them through `/settings/reset`.
(See e.g. [this
discussion](https://github.com/opentrons/opentrons/pull/17051#discussion_r1872275665),
though I sometimes have second thoughts about this.)

So we have been in a weird in-between where we were getting the dynamic
list of options from the server, but still making strong assumptions
about what that list would contain. This PR rewrites it to be fully
client-decided.

## Changelog

This is intended to not have any user-facing change.

The one "accidental" change that I'm aware of is that if the client is
newer than the server and has an extra reset option, it will now be
displayed, and filtered out of the HTTP request. Formerly, it would not
even be displayed.