feat(api): prepare for aspirate in air gap (#17919)
We should add a prepare-to-aspirate to the python protocol api `air_gap`
command, where it can issue the command after we've moved to a known-dry
position and before we actually dispense the airgap command. We know the
position is dry and therefore it's safe to do a prepare.

However, to be able to _unconditionally_ do a prepare, we also need to
add a fix to the engine prepare command to no-op if the pipette is
already prepared to aspirate. This is pretty important because if we've
already aspirated some liquid, we really don't want to move back to the
bottom position and dispense it all.


Useful testing protocol:
```
from opentrons.protocol_api import ProtocolContext

metadata = {"protocolName": "Test air gap prepare"}
requirements = {"apiLevel": "2.22", "robotType": "Flex"}


def run(protocol: ProtocolContext) -> None:
    plate = protocol.load_labware("biorad_384_wellplate_50ul", "A1")
    rack = protocol.load_labware("opentrons_flex_96_tiprack_50ul", "A2")
    pipette = protocol.load_instrument("flex_1channel_50", "left", tip_racks=[rack])

    pipette.pick_up_tip()
    pipette.aspirate(location=plate["A1"].bottom(z=2), volume=10)
    pipette.dispense(location=plate["A1"].top(), volume=10)
    pipette.blow_out(location=plate["A1"].top())
    pipette.air_gap(height=5, volume=10)

    pipette.dispense(location=plate["A2"].top(), volume=10)
    pipette.return_tip()
```


Closes EXEC-1391