fix(app): Fix Desktop LPC "apply offsets" state (#17940)
Closes RQA-4022

The "apply offsets" button disabled status utilized different state than the copy that would render the "offsets applied/not applied". These should be the same. More importantly, on the desktop, if a different user applied offsets, the missingSteps would not filter out LPC, since it was directly tied to a CTA.

Additionally, after talking to Design/QA, we've decided that when applying offsets, the "launch LPC" button should be disabled, too.