fix(api): dont crash subsystem manager with usage
The usage system just doesn't calculate its expected respondants
correctly. It would be sort of annoying to fix because a bunch of logic
would have to move around, and we don't really use the data for
anything, so let's just make it stop waiting for all the
responses (they'll show up in the logs anyway). That fixes an occasional
timeout that would crash the tool reader task.

Closes RQA-4020