fix(api): do not move plunger in 96chan DTW (#17943)
In the drop tip wizard, when you do an `unsafe/dropTipInPlace`, we were
telling the plunger to move to the bottom even if this was a cam action
setup. This isn't technically necessary, and if we do it then we will
cause an overpressure or stall to immediately reoccur which you really
don't want.

## testing
- you should be able to drop tips in the drop tip wizard using a
full-rack 96 config, even if you have an overpressure error that
persists into the drop tip wizard (like a clogged or physically deformed
tip)
- also support has some method for making the plunger stall which should
now be fixed as well

Closes RESC-418