feat(protocol-designer): preselect liquid class
This PR introduces logic to preselect a liquid class for a moveLiquid or mix step if all selected aspirate wells consist of the liquids with a single liquid class. If multiple liquid classes are included in the aspiration wells, or no liquid classes are included, then the form preselects "Don't use a liquid class". The logic for determining if a liquid class should be preselected and updating the form accordingly is contained in `useAssignLiquidClass`. This hook only sets the form liquidClass field if the form is presaved, so no saved step forms are affected. Whether or not the form is updated, the hook returns the assigned liquid class if any, to let the caller know how to order the liquid class options (assigned first).

Closes AUTH-1657