fix(app): Fix local state issues when "resetting to default" in LPC (#17965)
Closes RQA-4028 and RQA-4027

Fixes a bug in which the app would not delineate between "deleted offsets" and "updated offsets" when updating Redux state. In the saveWorkingOffsets command, after creating/deleting offsets, the server responses are added to a single array, so for example, "location C2 offset: 1,2,3". How do we know that this location + offset combo was updated to this value or this was the location + offset combo that was deleted? We don't, woops, so the client just assumed that all values were updated, and it didn't handle the delete case.