feat(protocol-designer): add designerApplication comment to py file a… (#17938)
…nd allow reupload

This PR adds the `DESIGNER_APPLICATION` variable which stringifies the
designer application key at the end of a python file. When you export,
the python file now has a long variable at the end with the string.

When you import back into PD, it now accepts the python file and
extracts the stringified blob back into a json.