chore(api): flex stacker platform sensing workaround (#17939)
# Overview
There are two platform sensors on the Flex Stacker X-track that allows
us to verify the shuttle is placed properly on the x-carriage. On the
EVT units, due to hardware limitations, the sensor on the extended side
of the track doesn't work. This means that we cannot reliably check for
the platform while the X-axis is extended.

This PR serves as a temporary fix and workaround until we receive new
hardware units by making the `flex_stacker.verify_shuttle_location`
function retracts the X-axis before taking a sensor reading.