fix(app): fix accumulating offsets on run record (#17969)
Closes EXEC-1411

When "cloning" a run on the app, the app quite literally creates a run with some predefined values gleaned from the historical run record of the run being cloned, and the new run reflects those old values. One of those values is labwareOffsets.

In the new LPC, a user clicks the "apply offsets" button, and this applies all the offsets utilized in the run to the run record. If a run is cloned, this mean that all the old offset values are present on the run record, and now the list increases by all the offsets utilized in this run. By cloning the run repeatedly, this list can grow unbounded.

This commit adds a filter when cloning a run so that only the most recent and unique labware offsets are preserved. In other words, the labware offsets field on the run record now can only ever be 2x the amount of unique offests possible for a run: the cloned run list of offsets + the offsets that are applied when clicking applied offsets.