feat(robot-server,system-server): Return server timing metrics in HTTP responses (#17970)
## Overview

When investigating server-side performance concerns, it's helpful to
isolate how much time is spent actually waiting for the server's Python
to chug away, versus how much time is spent in the network, or in nginx
or whatever.

This adds a FastAPI middleware to attach, to every HTTP response, a
header describing how much time our Python spent processing the request.
This is based on the standard
[Server-Timing](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Server-Timing)
header, so it automatically shows up in Chrome devtools.

Exactly what it measures is a little nuanced. It's an ASGI middleware,
so it ought to include FastAPI overhead. But the FastAPI docs note that
dependency teardown runs *after* the middleware, so 🤷.

Goes towards EXEC-1412.

## Test Plan and Hands on Testing

* [x] Push it to a robot, make sure the servers still boot and handle
requests, and test it in Chrome devtools.

<img width="1294" alt="Screenshot 2025-04-02 at 3 48 24 PM"
src="https://github.com/user-attachments/assets/b58d63d1-8391-488e-9f5f-974610b2e07e"
/>

## Changelog

* Implement the FastAPI middleware as described above.
* Use the middleware in robot-server and system-server.
* Update `server-utils`'s dev environment because something was broken
in the old dependencies and I couldn't use a `TestClient`.

## Review requests

OK with the vagueness of what's actually being measured?

## Risk assessment

Low.