fix(app): fix applying offsets implicitly when navigating on the desktop app (#17967)
Closes RQA-4032

Woops, the clientData effect for applying offsets for apps that did not do LPC had some logic issues. It was possible to navigate back and forth on the desktop app quickly and cause offsets to automatically apply, because some of the state used in the effect hadn't had the chance to properly initialize.

To fix, let's not update the client store as an effect. Since only one user can actually click the apply offsets button, that user should do the client store update on the CTA itself. This simplies the effect a bit and everything works.