fix(api): check flex stacker shuttle in the retracted location only for now (#17983)
# Overview
The previous workaround for stacker EVT introduced some motion-related
issues.

Currently, the `verify_shuttle_location` function always homes the
x-axis to the retracted position to check for shuttle presence. While
this approach technically works, it causes the motor driver to detect a
stall during the next move command, since the axis is already at the
limit switch.

A better solution is to eliminate this redundant movement from
`verify_shuttle_location`. Instead, we should perform the verification
only after the x-axis reaches the endstop as part of the normal dispense
and store routines.

It's annoying but these workarounds will be removed once we receive DVT
hardware!!