refactor(api): stop mocking protocol_api.validation in test_instrument_context.py (#17994)
# Overview

For some reason, we were mocking out the entire
`protocol_api/validation.py` module in `test_instrument_context.py`.
There is no reason to do that, since the validation functions are pure
functions that don't rely on external objects, and we really ought to
test the validation functions in our tests.

Furthermore, we were abusing the mocks to make the validation functions
do wild things: like mistranslating `Location`s into `WellTarget`s, or
turning the `"never"` tip-replacement policy into
`TransferTipPolicyV2.ONCE`. There is no good reason to do that, and it
just makes the test logic ridiculously hard to follow.

I need to fix this because I want to write a test for my change to the
`mix()` function, but it was very hard to do so given how we were using
`mock_validation`.

## Test Plan and Hands on Testing

I examined all the usages of `mock_validation` by hand. It was very
painful.

## Risk assessment

Low. This is a test-only change.