refactor(app-shell-odd): fix lint complaints that show up at the bottom of every PR (#18013)
# Overview

At the bottom of every PR in Github, it shows a bunch of lint errors for
`Unchanged files with check annotations`. This clutters up the code
review, and makes it hard to quickly find the bottom of the changed code
that I actually care about.

This PR fixes the lint errors:
- "Expected property shorthand"
- "Don't use `String` as a type. Use string instead"
- "This assertion is unnecessary since it does not change the type of
the expression"
- "Unexpected number value in conditional. An explicit zero/NaN check is
required"

## Test Plan and Hands on Testing

I'm relying on the existing CI tests.

## Risk assessment

Low. There should be no change in behavior. This is just a cleanup to
make lint happy.