feat(api): add the touch_tip(mm_from_edge=...) argument to PAPI (#17874)
# Overview

This adds the new argument `mm_from_edge` to the `touch_tip()` public
API function.

The option is needed for PD protocols. And we generally think that it
makes more sense for users to specify `mm_from_edge` instead of
`radius`.

Support for `mm_from_edge` already exists in the Protocol Core and
Engine (PR #17107), so all we need to do is to pass the new argument
along to the Core.

We follow the same convention that we adopted in the Protocol Engine,
that `radius` and `mm_from_edge` are mutually exclusive, so if the user
specifies `mm_from_edge`, they must leave `radius` to its default value
of 1.0

## Test Plan and Hands on Testing

I added unit tests to check for all combinations of `mm_from_edge` and
`radius`.

## Review requests

Do I need to do anything else for API versioning? What API version would
this go into?

## Risk assessment

Medium. We're adding a new API option, and we have to support API
changes forever once released.