refactor(app, step-generation): use `getLabwareDefIsStandard()` to check if labware is standard (#18029)
# Overview

We have a couple of places in the code where we say
`labwareDefinition.namespace === 'opentrons'` to check if something is
standard labware. But we have the shared-data function
`getLabwareDefIsStandard()` to do that, so let's use that everywhere.

(I'm doing this cleanup to prepare for custom labware in PD Python
export.)

## Test Plan and Hands on Testing

I'm relying on CI tests to make sure this works.

## Risk assessment

Low. There should be no functional change at all.