fix(api): Fix `TypeError` when the labware search path contains certain non-labware files (#18018)
This fixes RESC-428.

`opentrons_execute` and its friends automatically search certain
directories for files that look like labware definitions. They are
supposed to silently ignore files that are not labware definitions. The
bug is that if a file was a `.json` file containing something other than
an object, like `[]`, `"hello"`, `123`, or `null`, we would leak an
unhandled `TypeError` when we tried to read the `schemaVersion` property
from it.