feat(api): allow hardware modules listen for FLEX status bar update events (#18001)
# Overview
It's always been a pipe dream for some of us that one day the LEDs on
the connected hardware modules can align with the Opentrons robot's
status bar. We should make it happen!

This PR enables status bar state updates on the FLEX to become
listenable events and allows hardware modules to subscribe to relevant
API events. By leveraging event listeners, each module type gains the
flexibility to choose which events to respond to and handle them in a
way that best suits their needs.

We will most likely have to add firmware changes to other hardware
modules so this behavior will only be available for the FLEX Stacker for
now.