feat(step-generation): custom labware Python export (#18041)
# Overview

Add support for custom labware in PD Python export. Custom labware is
loaded with the API functions `load_labware_from_definition()` and
`load_adapter_from_definition()`.

Those functions take a **Python dict**, not a JSON string. But if we
emitted the labware definitions as Python dicts, PD would not be able to
use the labware definitions when importing the Python file, because PD
does not have a Python parser. (The biggest difference between Python
dicts and JSON is that Python uses `True`/`False`, whereas JSON uses
`true`/`false`.)

Our solution is that we will embed the custom labware definitions as
JSON strings in the Python file, and the Python code will call
`json.loads()` to parse the JSON into a Python dict. So our Python file
will look like this:

```
def run(protocol: protocol_api.ProtocolContext):
    well_plate_1 = protocol.load_labware_from_definition(
        CUSTOM_LABWARE["namespace/custom_labware/1"],
        location="B1",
        label="My labware",
    )

CUSTOM_LABWARE = json.loads("""
    {
        "namespace/custom_labware/1": { ... }
    }
""")
```

## Test Plan and Hands on Testing

Added unit tests.

I also tried it in PD with a piece of custom labware, and confirmed that
the generated protocol works in `simulate`.

## Risk assessment

Low-ish. Python export is behind a feature flag.