feat(abr-testing): Statically Clung Tips Removal Methods on 8ch Pipette (#18050)
<!--
Thanks for taking the time to open a Pull Request (PR)! Please make sure
you've read the "Opening Pull Requests" section of our Contributing
Guide:


https://github.com/Opentrons/opentrons/blob/edge/CONTRIBUTING.md#opening-pull-requests

GitHub provides robust markdown to format your PR. Links, diagrams,
pictures, and videos along with text formatting make it possible to
create a rich and informative PR. For more information on GitHub
markdown, see:


https://docs.github.com/en/get-started/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax

To ensure your code is reviewed quickly and thoroughly, please fill out
the sections below to the best of your ability!
-->

# Overview

Statically Clung Tips Removal Methods on 8ch Pipette

## Test Plan and Hands on Testing

- Tested in ABR

## Changelog

`trash_bin_tip_removal_.py` - This uses the left edge of the trash bin
with the 8ch pipette. The pipette drops the tip and swipes to the left
with the ejector down
`waste_chute_tip_removal.py` - This uses the left edge of the waste
chute with the 8ch in single or full tip pick up mode. The pipette drops
the tip and swipes to the left with the ejector down

## Review requests

<!--
- What do you need from reviewers to feel confident this PR is ready to
merge?
- Ask questions.
-->

## Risk assessment

These protocols do not analyze successfully. When they analyze, a
`TipNotAttachedError` is raised, but they run successfully on the robot.

---------

Co-authored-by: David Gomez <57112633+djgomez-opentrons@users.noreply.github.com>