perf(api,robot-server): Avoid blocking when logging to journald (#18075)
## Overview

This introduces a buffer so calls like `log.debug(...)` won't block when
journald has a hiccup and can't consume records quickly enough. This
implements one of the solutions discussed in #18034.

## Changelog

Calls like `log.debug(...)` are now sent to an in-memory,
global-to-the-process queue, so they will never block on I/O to
journald. In the background, a worker thread moves messages from the
queue to journald.

There are also the "sensor logs", which have the same workaround applied
to them, except it's logger->queue->`RotatingFileHandler` instead of
logger->queue->journald.

The ultimately emitted log records are intended to be unchanged, keeping
the same syslog IDs and such that we had before.