feat(api): PAPI: store labware objects in stacker
Adds PAPI bindings for the new stacker API.

While the engine commands handle all three of the (load the labware,
load the labware with these ids, move in the preexisting labware with
these ids) in each command - which I'm suspecting was a mistake - the
python api would be really ugly if we did that, so instead let's have a
separate version of fill and set_stored_labware to handle it. We also
now need to know how many labware we can hold because we won't silently
clamp lists of labware to inject, so we need to ask the user to do it.

The new api is therefore:
- set_stored_labware_items(), which takes a list of labware
- fill_items(), which takes a list of labware
- get_stored_labware(), which returns labware objects for all the
labware in the stacker
- get_max_storable_labware() and get_current_storable_labware(), which
give you the number of labware we can hold ever or right now
respectively, and fail if the stacker isn't configured
- get_max_storable_labware_from_list() and
get_current_storable_labware_from_list(), which are similar but instead
of returning a number and require configuration take a list and return a
list.

That last one is a doozy but is needed so that you can take the list of
labware you want to put in the stacker and pass it to that function and
save the result to pass in to fill_items() or
set_stored_labware_items().

There's also some implementation weirdness. The core implementation of
these functions needs to know about labware. That means that we need
access to labware type definitions, which means modulecore needs to be
generic on the labware like instrumentcore is, which means some things
need casts now; it also means that module cores need protocol core
backrefs, because some labware information access is done in the core as
protocol core methods rather than as labware core methods.

All in all somewhat chunky, but a nice API.

Closes EXEC-1397