fix(app): disable the run action button when uncurrenting a run (#18124)
Closes RQA-4118

On the desktop app, we implicitly uncurrent the run if the run has a cancelled run status. There's a brief window that's quite noticeable on the OT-2 in which the run has a "cancelled" run status but is still uncurrenting. If a user tries to create a new run during this window, the POST request fails due to server internals currently updating in light of the ongoing run uncurrenting.

To fix, block this intermediate state until the run is uncurrented.