fix(app): fix OT-2 LPC Offset Summary table displaying incorrect values (#18128)
Closes RQA-4126

In the updated OT-2 LPC, the offsets we display in the offset summary view are no longer a summation of the existing offset + the "working" offset, because by the time the user gets to the summary view, the existing offset has been updated on the server. We only need to consider the existing offset now when showing these values. This was just a visual bug: we are sending the correct offset values to the server.