refactor(app): handle new stacker errors
Nwo that the stacker retrieve command doesn't specify the labware ID we
have to get it a different way. We can do it from the errorInfo field on
the defined error, which is convenient. Unfortunately it's currently a
little inconvenient to pull that data out of the error types, so let's
refactor some of the utility code to make it a bit easier and a bit more
correct.