feat(api): store labware objects in the stacker (#18010)
We had this beautiful dream: the stacker wouldn't store labware.
Instead, the stacker would know what _kind_ of labware it should store,
and how many of those labware it is storing, and objects would be
created when you retrieved stuff. The problem with that is in the
physical world there are labware objects in the stacker, and if we're
not modeling them you can't know the state of those physical objects. So
we have to reverse this decision and store the identity of the labware
in the stacker.

This PR is the set of changes to accomplish that in the engine and
account for it in the python API; the ts bindings; and at least somewhat
in the app interface that already exists. It won't add any new API or
app stuff.

The way that this works is:
- The stacker contains a list of labware IDs that it contains
- While a labware is in the stacker, its position in a stack (i.e. lid
on labware, labware on adapter) is maintained but the stack as a whole
is `InStackerHopper(moduleId)`
- Labware must exist to go in to the stacker. That means that
`setStoredLabware` and `fill` now handle making sure all the labware
exists. They take lists of ids for PD compliance and also so that you
can predefine labware if you want. They also still have a count
argument. You can
- Say a number of labware with `count`, and the command will create that
many labware with arbitrary IDs
- Not set count, and pass a list of IDs that don't currently exist, and
the command will create labware with those IDs
- Not set count, and pass a list of IDs that do currently exist, and the
command will move those labware in
     - Those labware must be in `OFF_DECK` or there's an error
- Labware is moved out of the stacker, so `retrieve` doesn't create
labware anymore
- When you remove labware with `empty`, it's moved to `OFF_DECK`

`fill`, `setStoredLabware`, and `empty` all now specify lots of labware
location sequences and ids in their results because they create or move
labware where they previously do not. When `fill` and `setStoredLabware`
create labware, they gloss those labware as having an original position
descending to `SYSTEM`. Otherwise, they're the real original locations.

## to come out of draft
- [x] Fix the tests that already exist
- [x] Write new tests to cover extra verification of labware state for
the whole new-by-count/new-by-id/existing-id thing
- [x] Write new tests to make sure we do locations appropriately
- [ ] Test on hardware
- ~Make commit names that aren't so silly~ that's what the squash is for

## to merge
- [x] ~make a pr that implements this in the python api so it's
testable~ https://github.com/Opentrons/opentrons/pull/18108

## review requests
in addition to the standard stuff,
- Do you feel like too much/too little/just enough is in stacker common
- Is this an API the app can work with