refactor(shared-data): add constants for WellOrigins
Liquid classes contain references to PositionReferences in order to move within a well for submerging, aspirating, retracting, etc. However, protocol engine expects WellOrigins-- these are types that are distinct from PositionReferences, but map one-to-one with them. In order to supply protocol engine with the correct positional reference type, we need to map PositionReferences coming from PD to WellOrigins in our step-generation command creators.