refactor(app): add error handling for new stacker commands (#18133)
Now that the commands are changing, some of the error recovery had to
change too. Specifically `retrieve` doesn't have a labware ID param
anymore, which is what retrieve error-recovery used to display a
labware.

Instead, we can put that labware ID in the error info of the stacker
error objects, and then we can get that in error recovery. That takes
some typescript acrobatics to make sure that we can pull it out of the
right kind of error, and that is greatly aided by adding the defined
error types to the `RunTimeCommands`.

What I haven't done in this PR is add all the defined errors to all the
runtime commands, but we really really should: this is the point of
defined errors, that there's some errors where we know they can happen
on certain commands, and we should write that down in the client that we
write it down on the server.

## testing
- [x] do some retrieve error recovery passes and check that it functions
- [x] also check that command text works