feat(api): implement mix() parameters `aspirate_delay` and `dispense_delay` (#18000)
# Overview

In PD, you can specify a delay after each aspirate and dispense in a Mix
step. There is no way to specify a delay in the PAPI `mix()` call, which
means that we can't translate a PD Mix step into a PAPI `mix()` call.

This PR introduces the parameters `aspirate_delay` and `dispense_delay`
to the `InstrumentContext.mix()` function. The new parameters are gated
to API version 2.24 and above. AUTH-1366

## Test Plan and Hands on Testing

I added tests to demonstrate that this implementation emits the correct
calls to `delay()` when the caller requests a delay in the `mix()`.

This is what the output looks like in `simulate`:
```
Mixing 2 times with a volume of 20.0 ul
	Aspirating 20.0 uL from C2 of (Retired) Armadillo 96 Well Plate 200 µL PCR Full Skirt on slot A2 at 716.0 uL/sec
	Delaying for 0 minutes and 2.5 seconds
	Dispensing 20.0 uL into C2 of (Retired) Armadillo 96 Well Plate 200 µL PCR Full Skirt on slot A2 at 716.0 uL/sec
	...
```

## Review requests

`InstrumentContext.mix()` calls `InstrumentContext.aspirate()` and
`InstrumentContext.dispense()` to implement the mix, which is nice
because the public `aspirate()` and `dispense()` functions handle things
like publishing messages for the simulator.

Ideally, we would also call the public `delay()` function for the delay.
But the `delay()` function is in `ProtocolContext`, and
`InstrumentContext` doesn't have access to that. We only have access to
the `ProtocolCore`, so we have to publish messages to the simulator
manually.

I ended up refactoring the code of `mix()` a bit because it was going to
get too repetitive and nested with the delays and `publish_context()`
messages added.

## Risk assessment

Should be low. The new parameters are version-gated. The main risk is
that we release this, and then decide that we don't want to change the
public API this way.

(cherry picked from commit 501214809672045122a4370f2dad94c6bc99c260)