style(monorepo): enforce import order and styling with prettier (#18152)
In our JS code, we have a specific import order we like to observe, but it's currently not enforced programatically.

This commit adds import order enforcement programatically via prettier and cleans up all JS files. The changes that matter here are .prettierrc.js and the top-level package.json (for the added dev package for sorting import order by Regex). Everything else is just refactoring files to adhere to these rules.